+++ src/routes/directions.js
@@ -2,10 +2,10 @@ import UserRegistrationPage from '../views/userRegistrationPage/index'
 import CollaboratorRegistrationPage from '../views/collaboratorRegistrationPage/index'
 import HomePage from '../views/HomePage/index'
 import LoginPage from '../views/loginPage/index'
-import ExplorePage from '../views/explorePage/index'
-import ExplorePageTI from '../views/explorePageTI/index'
-import ExplorePageEncanador from '../views/explorePageEncanador/index'
-import ExplorePageJardineiro from '../views/explorePageJardineiro/index'
+import ExplorePage from '../views/explore/explorePage/index'
+import ExplorePageTI from '../views/explore/explorePageTI/index'
+import ExplorePageEncanador from '../views/explore/explorePageEncanador/index'
+import ExplorePageJardineiro from '../views/explore/explorePageJardineiro/index'
 
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
 import { AuthProvider } from "../context/auth"
