+++ src/views/explore/explorePageJardineiro/index.js
@@ -1,43 +1,43 @@
 import * as S from './styles'
-import Navbar from '../../components/Navbar/navbar'
-import Footer from '../../components/Footer/footer'
-import sinal from '../../assets/sinal.svg'
-import Button from '../../components/Button'
-import serviceTI from '../../assets/TiExample.png'
-import Diarista from '../../assets/Diarista.png'
-import Encanador from '../../assets/Encanador.png'
-import Pedreiro from '../../assets/pedreiro.png'
-import { api } from '../../services/api'
 import React, { useEffect, useState } from "react";
+import Navbar from '../../../components/Navbar/navbar'
+import Footer from '../../../components/Footer/footer'
+import sinal from '../../../assets/sinal.svg'
+import Button from '../../../components/Button'
+import serviceTI from '../../../assets/TiExample.png'
+import Diarista from '../../../assets/DiaristaExample.png'
+import Encanador from '../../../assets/EncadorExample.png'
+import Pedreiro from '../../../assets/Pedreiro.png'
+import { api } from '../../../services/api'
 
-function ExplorePageEncanador() {
 
+function ExplorePageJardineiro() {
   function ocuppation(e){
-    if (e===0){
+    if (e===1){
       return "Diarista"
     }
-    else if(e===1){
+    else if(e===2){
       return "Encanador"
     }
-    else if(e===2){
+    else if(e===3){
       return "Pedreiro"
     }
-    else{
+    else if (e==4){
       return "Tecnico"
     }
   }
 
   function image(e){
-    if (e===0){
+    if (e===1){
       return Diarista
     }
-    else if(e===1){
+    else if(e===2){
       return Encanador
     }
-    else if(e===2){
+    else if(e===3){
       return Pedreiro
     }
-    else{
+    else if (e==4){
       return serviceTI
     }
   }
@@ -56,12 +56,12 @@ function ExplorePageEncanador() {
   function listItems ()
   {
         return listItems= user?.map((number)  => 
-        {if(user?.[i++].occupation===1)
+        {if(user?.[i++].occupation===0)
         {
         return(
         <div>
           <img src={image(user[i-1]?.occupation)} alt=""></img>
-          <p id="titleService"> Nome do prestador: {user[i-1]?.name}</p>
+          <p id="titleService"> Prestador: {user[i-1]?.name}</p>
           <p id="descriptionService">Serviço: {ocuppation(user[i-1]?.occupation)}</p>
         </div>  
         )
@@ -69,6 +69,7 @@ function ExplorePageEncanador() {
         }
         );
   }
+
 
   return (
     <S.Container>
@@ -79,10 +80,10 @@ function ExplorePageEncanador() {
                 <div id="link">
                     <a href="/">Home</a>
                     <img src={sinal} alt="sinal maior"></img>
-                    <p id="service">Serviços de encanadores</p>
+                    <p id="service">Serviços de Jardinagem</p>
                 </div>
-                <p id="title">Serviços de Encanadores</p>
-                <p id="description">Aqui você pode encontrar os melhores encanadores.</p>
+                <p id="title">Serviços de Jardinagem</p>
+                <p id="description">Aqui você pode encontrar os melhores jardineiros.</p>
                 <div id="order-by">
                     <p id="order">Ordernar por:</p>
                     <Button
@@ -103,11 +104,10 @@ function ExplorePageEncanador() {
                 </div>
             </header>
         </div>
-
       </S.Body>     
       <Footer />
     </S.Container>
   )
 }
 
-export default ExplorePageEncanador
+export default ExplorePageJardineiro
